End-to-End training process using gluon and mxnet
In this assessment I will write a full end-to-end training process using gluon and mxnet. We will train the LeNet-5 classifier network on the MNIST dataset. The network will be defined for you but I have to fill in code to prepare the dataset, train the network, and evaluate it's performance on a held out dataset.